Use favicon executable to download favicon images
Start native `favicon` executable to download favicons for the listed
sites. This is slow, so runs is forked to the background. LaunchBar
will only see the image files that next time it runs. Images are saved
for at least 1 day, so won't be re-dlownloaded the next run. Old images
are continued to be used until they are replaced.